fix: sync wecom department order field issues

+++ backend/crm/src/main/java/cn/cordys/crm/integration/wecom/dto/WeComDepartment.java
@@ -44,7 +44,7 @@ public class WeComDepartment {
     /**
      * 在父部门中的次序值。order值大的排序靠前。值范围是[0, 2^32)
      */
-    private Integer order;
+    private Long order;
 
 
     /**
